Update the 3D examples to the last rust-nightly.
rustc 1.0.0-nightly (199bdcfef 2015-03-26).
The 2D examples are still broken.

+++ examples3/nphysics_testbed3d/src/objects/ball.rs
@@ -25,7 +25,7 @@ impl Ball {
             color:      color,
             base_color: color,
             delta:      delta,
-            gfx:        window.add_sphere(radius as f32),
+            gfx:        window.add_sphere(radius),
             body:       body
         };
